app/requirements.txt
django
djangorestframework
django-redis
gunicorn
